Few changes on the algorithm:
Server2 became Server and the old Server was deleted.

To run the Server just call the function initiate(). If weights values are passed it initializes the blocks with weights, and if not it is initialized with random values.

The weights are an array of integers like this:

  weights = [0,1,2,3,4,5,6]

  [0]--0--[0]--1--[0]
   |       |       |
   2       3       4
   |       |       |
  [0]--5--[0]--6--[0]

Now generates a file chose_name.txt with a description output of the blocks and a image file of the final QBT Block_index.png

It's merging by sorted values